+++ server/channels/api4/channel.go
@@ -6,15 +6,13 @@ package api4
 import (
 	"encoding/json"
 	"fmt"
-	"net/http"
-	"strconv"
-	"strings"
-	"time"
-
 	"github.com/mattermost/mattermost/server/public/model"
 	"github.com/mattermost/mattermost/server/public/shared/mlog"
 	"github.com/mattermost/mattermost/server/v8/channels/app"
 	"github.com/mattermost/mattermost/server/v8/channels/audit"
+	"net/http"
+	"strconv"
+	"strings"
 )
 
 func (api *API) InitChannel() {
@@ -81,8 +79,6 @@ func (api *API) InitChannel() {
 
 	api.BaseRoutes.ChannelModerations.Handle("", api.APISessionRequired(getChannelModerations)).Methods("GET")
 	api.BaseRoutes.ChannelModerations.Handle("/patch", api.APISessionRequired(patchChannelModerations)).Methods("PUT")
-
-	api.BaseRoutes.Channels.Handle("/admin-report", api.APISessionRequired(getChannelReport)).Methods("GET")
 }
 
 func createChannel(c *Context, w http.ResponseWriter, r *http.Request) {
@@ -2286,43 +2282,4 @@ func convertGroupMessageToChannel(c *Context, w http.ResponseWriter, r *http.Req
 	if err := json.NewEncoder(w).Encode(updatedChannel); err != nil {
 		c.Logger.Warn("Error while writing response from convertGroupMessageToChannel", mlog.Err(err))
 	}
-}
-
-func getChannelReport(c *Context, w http.ResponseWriter, r *http.Request) {
-	if !c.IsSystemAdmin() || !c.App.SessionHasPermissionTo(*c.AppContext.Session(), model.PermissionSysconsoleReadUserManagementChannels) {
-		c.SetPermissionError(model.PermissionSysconsoleReadUserManagementChannels)
-		return
-	}
-
-	sortColumn := "ChannelName"
-	if r.URL.Query().Get("sort_column") != "" {
-		sortColumn = r.URL.Query().Get("sort_column")
-	}
-
-	pageSize := 50
-	if pageSizeStr, err := strconv.ParseInt(r.URL.Query().Get("page_size"), 10, 64); err == nil {
-		pageSize = int(pageSizeStr)
-	}
-
-	filterOptions := &model.ChannelReportOptions{
-		model.ReportingBaseOptions{
-			SortColumn:          sortColumn,
-			SortDesc:            r.URL.Query().Get("sort_direction") == "desc",
-			PageSize:            pageSize,
-			LastSortColumnValue: r.URL.Query().Get("last_column_value"),
-			DateRange:           r.URL.Query().Get("date_range"),
-		},
-	}
-	filterOptions.PopulateDateRange(time.Now())
-
-	channelReports, appErr := c.App.GetChannelsReport(filterOptions)
-	if appErr != nil {
-		c.Err = appErr
-		return
-	}
-
-	if err := json.NewEncoder(w).Encode(channelReports); err != nil {
-		c.Logger.Error("Error writing channel report response", mlog.Err(err))
-	}
-
 }
