+++ server/channels/api4/user.go
@@ -108,8 +108,6 @@ func (api *API) InitUser() {
 
 	api.BaseRoutes.Users.Handle("/notify-admin", api.APISessionRequired(handleNotifyAdmin)).Methods("POST")
 	api.BaseRoutes.Users.Handle("/trigger-notify-admin-posts", api.APISessionRequired(handleTriggerNotifyAdminPosts)).Methods("POST")
-
-	api.BaseRoutes.Users.Handle("/admin-report", api.APISessionRequired(getUsersForReporting)).Methods("GET")
 }
 
 func createUser(c *Context, w http.ResponseWriter, r *http.Request) {
@@ -3443,62 +3441,5 @@ func getUsersWithInvalidEmails(c *Context, w http.ResponseWriter, r *http.Reques
 	err := json.NewEncoder(w).Encode(users)
 	if err != nil {
 		c.Logger.Warn("Error writing response", mlog.Err(err))
-	}
-}
-
-func getUsersForReporting(c *Context, w http.ResponseWriter, r *http.Request) {
-	if !(c.IsSystemAdmin() && c.App.SessionHasPermissionTo(*c.AppContext.Session(), model.PermissionSysconsoleReadUserManagementUsers)) {
-		c.SetPermissionError(model.PermissionSysconsoleReadUserManagementUsers)
-		return
-	}
-
-	sortColumn := "Username"
-	if r.URL.Query().Get("sort_column") != "" {
-		sortColumn = r.URL.Query().Get("sort_column")
-	}
-
-	pageSize := 50
-	if pageSizeStr, err := strconv.ParseInt(r.URL.Query().Get("page_size"), 10, 64); err == nil {
-		pageSize = int(pageSizeStr)
-	}
-
-	teamFilter := r.URL.Query().Get("team_filter")
-	if !(teamFilter == "" || model.IsValidId(teamFilter)) {
-		c.Err = model.NewAppError("getUsersForReporting", "api.getUsersForReporting.invalid_team_filter", nil, "", http.StatusBadRequest)
-		return
-	}
-
-	hideActive := r.URL.Query().Get("hide_active") == "true"
-	hideInactive := r.URL.Query().Get("hide_inactive") == "true"
-	if hideActive && hideInactive {
-		c.Err = model.NewAppError("getUsersForReporting", "api.getUsersForReporting.invalid_active_filter", nil, "", http.StatusBadRequest)
-		return
-	}
-
-	options := &model.UserReportOptions{
-		ReportingBaseOptions: model.ReportingBaseOptions{
-			SortColumn:          sortColumn,
-			SortDesc:            r.URL.Query().Get("sort_direction") == "desc",
-			PageSize:            pageSize,
-			LastSortColumnValue: r.URL.Query().Get("last_column_value"),
-			DateRange:           r.URL.Query().Get("date_range"),
-		},
-		Team:         teamFilter,
-		LastUserId:   r.URL.Query().Get("last_id"),
-		Role:         r.URL.Query().Get("role_filter"),
-		HasNoTeam:    r.URL.Query().Get("has_no_team") == "true",
-		HideActive:   hideActive,
-		HideInactive: hideInactive,
-	}
-	options.PopulateDateRange(time.Now())
-
-	userReports, err := c.App.GetUsersForReporting(options)
-	if err != nil {
-		c.Err = err
-		return
-	}
-
-	if jsonErr := json.NewEncoder(w).Encode(userReports); jsonErr != nil {
-		c.Logger.Warn("Error writing response", mlog.Err(jsonErr))
 	}
 }
