+++ server/public/model/client4.go
@@ -166,6 +166,10 @@ func (c *Client4) usersRoute() string {
 	return "/users"
 }
 
+func (c *Client4) reportsRoute() string {
+	return "/reports"
+}
+
 func (c *Client4) userRoute(userId string) string {
 	return fmt.Sprintf(c.usersRoute()+"/%v", userId)
 }
@@ -1954,7 +1958,7 @@ func (c *Client4) GetUsersForReporting(ctx context.Context, options *UserReportO
 		values.Set("date_range", options.DateRange)
 	}
 
-	r, err := c.DoAPIGet(ctx, c.usersRoute()+"/report?"+values.Encode(), "")
+	r, err := c.DoAPIGet(ctx, c.reportsRoute()+"/users?"+values.Encode(), "")
 	if err != nil {
 		return nil, BuildResponse(r), err
 	}
